epochmax and inputs output
Added variable, "epochmax", which makes it slightly easier to change how many epochs the program will run for.
     This variable also changes the center of mass kept in mutations.
Added a "center of mass" variable that changes after "epochmax/2" generations. The "center of mass" is branched off of for new mutants. The idea is that after some point the mixing algorithm will get close enough to the right location that we can just check around a region with the mutants.
   The current heatmap in the "cost" method is of a 2x2 square slightly above the middle of the FEM mesh. So far the program has located this fault within 1000 epochs.